+++ models/MessageEntry.php
@@ -3,6 +3,7 @@
 namespace humhub\modules\mail\models;
 
 use DateTime;
+use humhub\interfaces\ViewableInterface;
 use humhub\modules\user\models\User;
 use Yii;
 
@@ -12,7 +13,7 @@ use Yii;
  * @package humhub.modules.mail.models
  * @since 0.5
  */
-class MessageEntry extends AbstractMessageEntry
+class MessageEntry extends AbstractMessageEntry implements ViewableInterface
 {
     /**
      * @inheritdoc
@@ -60,5 +61,15 @@ class MessageEntry extends AbstractMessageEntry
             ->andWhere(['!=', 'id', $this->id])
             ->andWhere(['>=', 'created_at', $today])
             ->exists();
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function canView($user = null): bool
+    {
+        $message = $this->message;
+
+        return $message instanceof Message && $message->isParticipant($user);
     }
 }